requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
pyyaml==6.0.1
imap-tools==1.5.0
groq>=0.9.0
beautifulsoup4==4.12.3
openai>=1.30.0
watchdog
google-generativeai>=0.8.0
anthropic>=0.40.0
